% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrants.R
\name{assign_quadrants}
\alias{assign_quadrants}
\title{Assign SwitchClass quadrants}
\usage{
assign_quadrants(delta, fc, delta_thresh = 0.15)
}
\arguments{
\item{delta}{named numeric vector (features).}

\item{fc}{named numeric vector (same feature names as \code{delta}).}

\item{delta_thresh}{numeric threshold for |delta|.}
}
\value{
factor vector of quadrant labels with names = features.
}
\description{
Q1:  delta >  T & FC >  0
Q2:  delta < -T & FC >  0
Q3:  delta < -T & FC <  0
Q4:  delta >  T & FC <  0
Q0:  otherwise
}
